roll back, added astype(int) to accommodate windows machine users, removed unused import

+++ tests/test_lm.py
@@ -421,7 +421,7 @@ def test_using_int_as_intercept():
         0
     )
     assert isinstance(eq, LinearModel)
-    pred = eq.predict(EXAMPLE_DF)
+    pred = eq.predict(EXAMPLE_DF).astype(int)
     assert isinstance(pred, pd.Series)
     assert pred.dtype == int
     assert (pred.index == EXAMPLE_DF.index).all()
